refactor(case): split nodes.py into nodes/ subpackage (#736)
Split the 1502-line monolithic vultron/core/behaviors/case/nodes.py into a
nodes/ subpackage organized by semantic domain:

- conditions.py: CheckCaseAlreadyExists, CheckCaseExistsForReport,
  ValidateCaseObject
- case_setup.py: PersistCase, SetCaseAttributedTo, RecordCaseCreationEvents,
  CreateCaseActorNode
- participant.py: _create_and_attach_participant helpers +
  CreateCaseOwnerParticipant, CreateCaseParticipantNode
- embargo.py: _DEFAULT_EMBARGO_DAYS, _preferred_embargo_duration,
  _activate_default_embargo, InitializeDefaultEmbargoNode
- communication.py: EmitCreateCaseActivity, SendOfferCaseManagerRoleNode
- lifecycle.py: CommitCaseLogEntryNode
- __init__.py: re-exports all 13 classes + UpdateActorOutbox +
  _create_and_attach_participant (backward-compatible public path)

Tests migrated from test_nodes.py and test_commit_log_entry_node.py to
matching submodule test files. Patch paths updated in test_lifecycle.py
to point at the new module location. Added real unit tests for condition
nodes and InitializeDefaultEmbargoNode.

No behavior changes; all existing import paths continue to work via __init__.py.

Closes #736
