+++ vultron/core/behaviors/case/nodes/conditions.py
@@ -1,0 +1,166 @@
+#!/usr/bin/env python
+
+#  Copyright (c) 2026 Carnegie Mellon University and Contributors.
+#  - see Contributors.md for a full list of Contributors
+#  - see ContributionInstructions.md for information on how you can Contribute to this project
+#  Vultron Multiparty Coordinated Vulnerability Disclosure Protocol Prototype is
+#  licensed under a MIT (SEI)-style license, please see LICENSE.md distributed
+#  with this Software or contact permission@sei.cmu.edu for full terms.
+#  Created, in part, with funding and support from the United States Government
+#  (see Acknowledgments file). This program may include and/or can make use of
+#  certain third party source code, object code, documentation and other files
+#  ("Third Party Software"). See LICENSE.md for more details.
+#  Carnegie Mellon®, CERT® and CERT Coordination Center® are registered in the
+#  U.S. Patent and Trademark Office by Carnegie Mellon University
+
+"""
+Condition nodes for case management behavior trees.
+
+Provides idempotency guard conditions for the create_case workflow.
+Per specs/idempotency.yaml ID-04-004.
+"""
+
+from py_trees.common import Status
+
+from vultron.core.behaviors.helpers import DataLayerCondition
+from vultron.core.models.vultron_types import VultronCase
+
+
+class CheckCaseAlreadyExists(DataLayerCondition):
+    """
+    Check if a VulnerabilityCase already exists in DataLayer.
+
+    Returns SUCCESS if the case already exists (idempotency early exit).
+    Returns FAILURE if the case does not exist (proceed with creation).
+
+    Per specs/idempotency.yaml ID-04-004.
+    """
+
+    def __init__(self, case_id: str, name: str | None = None):
+        super().__init__(name=name or self.__class__.__name__)
+        self.case_id = case_id
+
+    def update(self) -> Status:
+        if self.datalayer is None:
+            self.logger.error(f"{self.name}: DataLayer not available")
+            return Status.FAILURE
+
+        try:
+            existing = self.datalayer.read(self.case_id)
+            if existing is None:
+                self.logger.debug(
+                    f"{self.name}: Case {self.case_id} not found, proceeding"
+                )
+                return Status.FAILURE
+
+            # A case record that exists but has no participants was
+            # pre-stored by the inbox endpoint as a dehydrated reference.
+            # It still needs full initialisation (participant creation,
+            # CaseActor setup, etc.), so return FAILURE to let the
+            # CreateCaseFlow run.
+            participants = getattr(existing, "case_participants", None) or []
+            if not participants:
+                self.logger.debug(
+                    f"{self.name}: Case {self.case_id} exists but has no"
+                    " participants — proceeding with initialisation"
+                )
+                return Status.FAILURE
+
+            self.logger.info(
+                f"{self.name}: Case {self.case_id} already exists"
+                " — skipping creation (idempotent)"
+            )
+            return Status.SUCCESS
+
+        except Exception as e:
+            self.logger.error(
+                f"{self.name}: Error checking case existence: {e}"
+            )
+            return Status.FAILURE
+
+
+class CheckCaseExistsForReport(DataLayerCondition):
+    """
+    Check if a VulnerabilityCase already exists for the given report.
+
+    Uses ``find_case_by_report_id`` to check whether a case linked to the
+    report already exists with participants. Returns SUCCESS if so (idempotency
+    early exit), FAILURE otherwise.
+
+    Per specs/idempotency.yaml ID-04-004.
+    """
+
+    def __init__(self, report_id: str, name: str | None = None):
+        super().__init__(name=name or self.__class__.__name__)
+        self.report_id = report_id
+
+    def update(self) -> Status:
+        if self.datalayer is None:
+            self.logger.error(f"{self.name}: DataLayer not available")
+            return Status.FAILURE
+
+        try:
+            existing = self.datalayer.find_case_by_report_id(self.report_id)
+            if existing is None:
+                self.logger.debug(
+                    f"{self.name}: No case found for report {self.report_id}"
+                )
+                return Status.FAILURE
+
+            # A case record that exists but has no participants was
+            # pre-stored by the inbox endpoint as a dehydrated reference.
+            # It still needs full initialisation, so return FAILURE.
+            participants = getattr(existing, "case_participants", None) or []
+            if not participants:
+                self.logger.debug(
+                    f"{self.name}: Case for report {self.report_id} exists"
+                    " but has no participants — proceeding with initialisation"
+                )
+                return Status.FAILURE
+
+            self.logger.info(
+                "%s: Case %s already exists for report %s"
+                " — skipping (idempotent)",
+                self.name,
+                existing.id_,
+                self.report_id,
+            )
+            return Status.SUCCESS
+
+        except Exception as e:
+            self.logger.error(
+                f"{self.name}: Error checking case existence: {e}"
+            )
+            return Status.FAILURE
+
+
+class ValidateCaseObject(DataLayerCondition):
+    """
+    Validate the incoming VulnerabilityCase object has required fields.
+
+    Returns SUCCESS if the case object passes validation.
+    Returns FAILURE if required fields are missing.
+    """
+
+    def __init__(self, case_obj: VultronCase, name: str | None = None):
+        super().__init__(name=name or self.__class__.__name__)
+        self.case_obj = case_obj
+
+    def update(self) -> Status:
+        try:
+            if self.case_obj is None:
+                self.logger.error(f"{self.name}: case_obj is None")
+                return Status.FAILURE
+
+            if not getattr(self.case_obj, "id_", None):
+                self.logger.error(f"{self.name}: Case object missing id_")
+                return Status.FAILURE
+
+            self.logger.debug(
+                f"{self.name}: Case {self.case_obj.id_} passes validation"
+            )
+            return Status.SUCCESS
+
+        except Exception as e:
+            self.logger.error(f"{self.name}: Error validating case: {e}")
+            return Status.FAILURE
